requirements.txt
# Core runtime dependencies
loguru>=0.7,<1
bleak>=0.21,<1
python-kasa>=0.6.2

# Development and testing
pytest>=7,<8
